+++ src/main/java/utility/ViewManager.java
@@ -43,12 +43,9 @@ public class ViewManager {
         }
 
         // For other event types, broadcast to components
-        components.stream()
-                .filter(component -> event.getTypes().stream().anyMatch(component::supportsEvent))
-                .forEach(component -> {
-                    System.out.println("Broadcasting " + event.getClass().getSimpleName() + " to " + component.getClass().getSimpleName());
-                    component.receiveViewEvent(event);
-                });
+        components.forEach(component -> {
+            component.receiveViewEvent(event);
+        });
     }
 
     private void switchPanel(String panelName) {