+++ src/main/java/utility/exceptions/RateLimitExceededException.java
@@ -1,0 +1,4 @@
+package utility.exceptions;
+
+public class RateLimitExceededException extends Exception {
+}